requirements.txt
pyspark
requests
pyyaml
delta-spark